backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
torch
torchvision
pillow
numpy
opencv-python
